requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
flask-restx==0.3.0
python-dotenv==0.17.0